requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
cffi==1.15.1
Django==4.1.7
django-environ==0.9.0
djangorestframework==3.14.0
Pillow==9.4.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
